Move get_local_ips() and is_file_local() to client.py
